Complete super cal

+++ FunSuper/FunSuper/Server/Entities/Disbursement.cs
@@ -5,7 +5,7 @@ namespace FunSuper.Server.Entities
     public class Disbursement : Entity
     {
         public int DisbursementId { get; set; }
-        public double SgcAmount { get; set; }
+        public decimal SgcAmount { get; set; }
         public DateTime PayMadeDate { get; set; }
         public DateTime PayFromDate { get; set; }
         public DateTime PayToDate { get; set; }
